art_gallery_backend/requirements.txt
flask_cors
pymongo
python-dotenv
